+++ py/desisurvey/test/test_utils.py
@@ -51,15 +51,13 @@ class TestUtils(unittest.TestCase):
     def tearDown(self):
         utils._iers_is_frozen = False
 
-    def test_dome_probs(self):
-        """Checks of dome-closed probabilities"""
-        probs = utils.dome_closed_probabilities()
-        self.assertEqual(len(probs), 12)
-        self.assertTrue(np.all(probs > 0))
-        self.assertTrue(np.all(probs < 1))
-        self.assertTrue(np.all(probs > 0))
-        self.assertTrue(probs[6] > 0.5) # July > 50%
-        self.assertTrue(probs[5] < 0.2) # June < 20%
+    def test_dome_fracs(self):
+        """Checks of dome-closed fractions"""
+        probs = utils.dome_closed_fractions()
+        self.assertTrue(np.all(probs >= 0))
+        self.assertTrue(np.all(probs <= 1))
+        mean = probs.mean()
+        self.assertTrue((mean > 0.25) & (mean < 0.35))
 
     def test_update_iers_bad_ext(self):
         """Test save_name extension check"""
@@ -277,12 +275,16 @@ class TestUtils(unittest.TestCase):
             100, utils.day_number(first + datetime.timedelta(days=100)))
 
     def test_monsoon(self):
-        """Monsoon based on (month, day) comparisons"""
-        for year in range(2019, 2025):
-            self.assertFalse(utils.is_monsoon(datetime.date(year, 7, 12)))
-            self.assertTrue(utils.is_monsoon(datetime.date(year, 7, 13)))
-            self.assertTrue(utils.is_monsoon(datetime.date(year, 8, 26)))
-            self.assertFalse(utils.is_monsoon(datetime.date(year, 8, 27)))
+        """Sanity checks on monsoon calculations"""
+        for year in range(2020, 2025):
+            self.assertFalse(utils.is_monsoon(datetime.date(year, 1, 1)))
+            self.assertFalse(utils.is_monsoon(datetime.date(year, 7, 11)))
+            self.assertFalse(utils.is_monsoon(datetime.date(year, 8, 20)))
+            self.assertFalse(utils.is_monsoon(datetime.date(year, 12, 31)))
+        self.assertTrue(utils.is_monsoon(datetime.date(2020, 8, 1)))
+        self.assertFalse(utils.is_monsoon(datetime.date(2021, 8, 1)))
+        self.assertTrue(utils.is_monsoon(datetime.date(2022, 8, 1)))
+        self.assertTrue(utils.is_monsoon(datetime.date(2023, 8, 1)))
 
     def test_local_noon(self):
         """The telescope is 7 hours behind of UTC during winter and summer."""